backend/requirements.txt
# Core
fastapi==0.115.5
uvicorn[standard]==0.32.1
python-multipart==0.0.12

# Database
sqlalchemy==2.0.36
greenlet==3.1.1
alembic==1.14.0
asyncpg==0.30.0
psycopg2-binary==2.9.10

# Auth
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1

# Validation
pydantic==2.10.3
pydantic-settings==2.6.1
email-validator==2.2.0

# Barcode / QR
python-barcode==0.15.1
qrcode[pil]==8.0
Pillow==11.0.0
reportlab==4.2.5

# Import pipeline
pandas==2.2.3
openpyxl==3.1.5
xlrd==2.0.1

# AI / ML
scikit-learn==1.5.2
numpy==2.1.3
scipy==1.14.1

# HTTP client (for AI services)
httpx==0.28.0

# Utilities
python-slugify==8.0.4
python-dotenv==1.0.1
structlog==24.4.0
tenacity==9.0.0

# MQTT (phase 2 ready)
paho-mqtt==2.1.0

# Testing
pytest==8.3.4
pytest-asyncio==0.24.0
httpx==0.28.0
